Change double to long double & fix potential issues

+++ src/util/QGramVector.cpp
@@ -14,12 +14,12 @@ QGramVector::QGramVector(std::string_view const& string, unsigned q) {
     CalculateLength();
 }
 
-double QGramVector::InnerProduct(QGramVector const& other) const {
+long double QGramVector::InnerProduct(QGramVector const& other) const {
     auto const& q_grams_pair = std::minmax(
         q_grams_, other.q_grams_, [](auto const& a, auto const& b) { return a.size() < b.size(); });
     auto const& l = q_grams_pair.first;
     auto const& r = q_grams_pair.second;
-    return std::accumulate(l.cbegin(), l.cend(), 0, [&r](double a, auto const& b) {
+    return std::accumulate(l.cbegin(), l.cend(), 0.0, [&r](double a, auto const& b) {
         auto r_pair = r.find(b.first);
         if (r_pair == r.end()) return a;
         return a + b.second * r_pair->second;
@@ -28,7 +28,7 @@ double QGramVector::InnerProduct(QGramVector const& other) const {
 
 void QGramVector::CalculateLength() {
     length_ = std::sqrt(
-        std::accumulate(q_grams_.cbegin(), q_grams_.cend(), 0, [](double a, auto const& pair) {
+        std::accumulate(q_grams_.cbegin(), q_grams_.cend(), 0.0, [](double a, auto const& pair) {
             return a + pair.second * pair.second;
         }));
 }
